+++ pypeit/spectrographs/__init__.py
@@ -23,3 +23,4 @@ from pypeit.spectrographs import lbt_luci
 from pypeit.spectrographs import mmt_binospec
 from pypeit.spectrographs import not_alfosc
 from pypeit.spectrographs import p200_dbsp
+from pypeit.spectrographs import mmt_mmirs